Adds Parameters to profiles.json
Dashboard.js now expects the profile.parameters to be passed in as
props. The file profiles.json did not define these values so when App.js
passed the selected profile to Dashboard using profiles.json data, a
JavaScript error was thrown and the page looked to be stuck in a loading
state.

This commit fixes the issue by adding the profile parameters to
profiles.json.